add user repository interface
