+++ src/Authorization/PL.Authorization/User/IUserRepository.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PL.Authorization.Application.User
+{
+    public interface IUserRepository
+    {
+        Task AddAsync(User user);
+    }
+}
